switch precision to double

+++ large_scale_hts_reconciliation/include/G.h
@@ -16,32 +16,32 @@
 #include <Eigen/SparseLU>
 #include <Eigen/SparseQR>
 
-typedef Eigen::SparseMatrix<float, Eigen::ColMajor> SpMat;
-typedef Eigen::Triplet<float> T;
+typedef Eigen::SparseMatrix<double, Eigen::ColMajor> SpMat;
+typedef Eigen::Triplet<double> T;
 
 using namespace Eigen;
 namespace lhts {
 namespace G {
 SpMat build_sparse_OLS(SpMat S);
-SpMat build_sparse_WLS(SpMat S, float w);
-SpMat build_sparse_top_down(const MatrixXi S_compact, const MatrixXf P,
+SpMat build_sparse_WLS(SpMat S, double w);
+SpMat build_sparse_top_down(const MatrixXi S_compact, const MatrixXd P,
                             int num_leaves, int num_nodes, int num_levels);
-SpMat build_sparse_middle_out(const MatrixXi S_compact, const MatrixXf P,
+SpMat build_sparse_middle_out(const MatrixXi S_compact, const MatrixXd P,
                               int level, int num_leaves, int num_nodes,
                               int num_levels);
 
 SpMat build_sparse_bottom_up(const MatrixXi S_compact, int num_leaves,
                              int num_nodes, int num_levels);
-MatrixXf build_dense_OLS(const MatrixXi S);
-MatrixXf build_dense_WLS(const MatrixXi S, float w);
-MatrixXf build_dense_top_down(const MatrixXi S_compact,
-                                     const MatrixXf P, int num_leaves,
+MatrixXd build_dense_OLS(const MatrixXi S);
+MatrixXd build_dense_WLS(const MatrixXi S, double w);
+MatrixXd build_dense_top_down(const MatrixXi S_compact,
+                                     const MatrixXd P, int num_leaves,
                                      int num_nodes, int num_levels);
-MatrixXf build_dense_middle_out(const MatrixXi S_compact,
-                                       const MatrixXf P, int level,
+MatrixXd build_dense_middle_out(const MatrixXi S_compact,
+                                       const MatrixXd P, int level,
                                        int num_leaves, int num_nodes,
                                        int num_levels);
-MatrixXf build_dense_bottom_up(const MatrixXi S_compact,
+MatrixXd build_dense_bottom_up(const MatrixXi S_compact,
                                       int num_leaves, int num_nodes,
                                       int num_levels);
 }  // namespace G
