+++ large_scale_hts_reconciliation/include/S.h
@@ -16,8 +16,8 @@
 #include <Eigen/SparseLU>
 #include <Eigen/SparseQR>
 
-typedef Eigen::SparseMatrix<float, Eigen::ColMajor> SpMat;
-typedef Eigen::Triplet<float> T;
+typedef Eigen::SparseMatrix<double, Eigen::ColMajor> SpMat;
+typedef Eigen::Triplet<double> T;
 
 using namespace Eigen;
 namespace lhts {
